+++ tests/app/settings.py
@@ -16,6 +16,7 @@ INSTALLED_APPS = [
     'wagtail.snippets',
     'wagtail.images',
     'wagtail.documents',
+    'wagtail.contrib.modeladmin',
 
     'django.contrib.admin',
     'django.contrib.auth',