+++ wagtailvideos/__init__.py
@@ -4,6 +4,11 @@ from django.core.exceptions import ImproperlyConfigured
 default_app_config = 'wagtailvideos.apps.WagtailVideosApp'
 
 
+def is_modeladmin_installed():
+    from django.apps import apps
+    return apps.is_installed('wagtail.contrib.modeladmin')
+
+
 def get_video_model_string():
     return getattr(settings, 'WAGTAILVIDEOS_VIDEO_MODEL', 'wagtailvideos.Video')