+++ wagtailvideos/models.py
@@ -6,28 +6,27 @@ import shutil
 import subprocess
 import tempfile
 import threading
-from contextlib import contextmanager
 from distutils.version import LooseVersion
 
+import bcp47
 import wagtail
 from django.conf import settings
 from django.core.exceptions import SuspiciousFileOperation
 from django.core.files.base import ContentFile
-from django.core.files.temp import NamedTemporaryFile
 from django.db import models
-from django.db.models.signals import post_save, pre_delete
-from django.dispatch.dispatcher import receiver
 from django.forms.utils import flatatt
 from django.urls import reverse
 from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_lazy as _
 from enumchoicefield import ChoiceEnum, EnumChoiceField
+from modelcluster.fields import ParentalKey
+from modelcluster.models import ClusterableModel
 from taggit.managers import TaggableManager
-from wagtail.core.models import CollectionMember
+from wagtail.core.models import CollectionMember, Orderable
 from wagtail.search import index
 from wagtail.search.queryset import SearchableQuerySetMixin
 
-from wagtailvideos import ffmpeg
+from wagtailvideos import get_video_model_string
 
 if LooseVersion(wagtail.__version__) >= LooseVersion('2.7'):
     from wagtail.admin.models import get_object_usage
@@ -180,13 +179,6 @@ class AbstractVideo(CollectionMember, index.Indexed, models.Model):
     def get_transcode_model(cls):
         return cls.transcodes.rel.related_model
 
-    def get_transcode(self, media_format):
-        Transcode = self.get_transcode_model()
-        try:
-            return self.transcodes.get(media_format=media_format)
-        except Transcode.DoesNotExist:
-            return self.do_transcode(media_format)
-
     def video_tag(self, attrs=None):
         if attrs is None:
             attrs = {}
@@ -205,14 +197,19 @@ class AbstractVideo(CollectionMember, index.Indexed, models.Model):
                        .format(self.url, mime.guess_type(self.url)[0]))
 
         sources.append("<p>Sorry, your browser doesn't support playback for this video</p>")
+
+        tracks = []
+        if hasattr(self, 'track_listing'):
+            tracks = [t.track_tag() for t in self.track_listing.tracks.all()]
+
         return mark_safe(
-            "<video {0}>\n{1}\n</video>".format(flatatt(attrs), "\n".join(sources)))
+            "<video {0}>\n{1}\n{2}\n</video>".format(flatatt(attrs), "\n".join(sources), "\n".join(tracks)))
 
     def do_transcode(self, media_format, quality):
         transcode, created = self.transcodes.get_or_create(
             media_format=media_format,
         )
-        if transcode.processing is False:
+        if created or transcode.processing is False:
             transcode.processing = True
             transcode.error_messages = ''
             transcode.quality = quality
@@ -292,61 +289,6 @@ class TranscodingThread(threading.Thread):
             shutil.rmtree(output_dir, ignore_errors=True)
 
 
-@contextmanager
-def get_local_file(file):
-    """
-    Get a local version of the file, downloading it from the remote storage if
-    required. The returned value should be used as a context manager to
-    ensure any temporary files are cleaned up afterwards.
-    """
-    try:
-        with open(file.path):
-            yield file.path
-    except NotImplementedError:
-        _, ext = os.path.splitext(file.name)
-        with NamedTemporaryFile(prefix='wagtailvideo-', suffix=ext) as tmp:
-            try:
-                file.open('rb')
-                for chunk in file.chunks():
-                    tmp.write(chunk)
-            finally:
-                file.close()
-            tmp.flush()
-            yield tmp.name
-
-
-# Delete files when model is deleted
-@receiver(pre_delete, sender=Video)
-def video_delete(sender, instance, **kwargs):
-    instance.thumbnail.delete(False)
-    instance.file.delete(False)
-
-
-# Fields that need the actual video file to create
-@receiver(post_save, sender=Video)
-def video_saved(sender, instance, **kwargs):
-    if not ffmpeg.installed():
-        return
-
-    if hasattr(instance, '_from_signal'):
-        return
-
-    has_changed = instance._initial_file is not instance.file
-    filled_out = instance.thumbnail is not None and instance.duration is not None
-    if has_changed or not filled_out:
-        with get_local_file(instance.file) as file_path:
-            if has_changed or instance.thumbnail is None:
-                instance.thumbnail = ffmpeg.get_thumbnail(file_path)
-
-            if has_changed or instance.duration is None:
-                instance.duration = ffmpeg.get_duration(file_path)
-
-    instance.file_size = instance.file.size
-    instance._from_signal = True
-    instance.save()
-    del instance._from_signal
-
-
 class AbstractVideoTranscode(models.Model):
     media_format = EnumChoiceField(MediaFormats)
     quality = EnumChoiceField(VideoQuality, default=VideoQuality.default)
@@ -377,7 +319,59 @@ class VideoTranscode(AbstractVideoTranscode):
         )
 
 
-# Delete files when model is deleted
-@receiver(pre_delete, sender=VideoTranscode)
-def transcode_delete(sender, instance, **kwargs):
-    instance.file.delete(False)
+class TrackListing(ClusterableModel):
+    video = models.OneToOneField(
+        get_video_model_string(), on_delete=models.CASCADE,
+        related_name='track_listing')
+
+    def __str__(self):
+        return self.video.title
+
+
+class VideoTrack(Orderable):
+    listing = ParentalKey(TrackListing, related_name='tracks', on_delete=models.CASCADE)
+    # TODO move to TextChoices once django < 3 is dropped
+    track_kinds = [
+        ('subtitles', 'Subtitles'),
+        ('captions', 'Captions'),
+        ('descriptions', 'Descriptions'),
+        ('chapters', 'Chapters'),
+        ('metadata', 'Metadata'),
+    ]
+
+    file = models.FileField(
+        verbose_name=_('file'),
+        upload_to=get_upload_to
+    )
+    kind = models.CharField(max_length=50, choices=track_kinds, default=track_kinds[0][0])
+    label = models.CharField(
+        max_length=255, blank=True,
+        help_text='A user-readable title of the text track.')
+    language = models.CharField(
+        max_length=50,
+        choices=[(v, k) for k, v in bcp47.languages.items()],
+        default='en', blank=True, help_text='Required if type is "Subtitle"', unique=True)
+
+    def track_tag(self):
+        attrs = {
+            'kind': self.kind,
+            'src': self.url,
+        }
+        if self.label:
+            attrs['label'] = self.label
+        if self.language:
+            attrs['srclang'] = self.language
+
+        return "<track {0}{1}>".format(flatatt(attrs), ' default' if self.sort_order == 0 else '')
+
+    def __str__(self):
+        return "{0} - {1}".format(self.label or self.get_kind_display(), self.get_language_display())
+
+    @property
+    def url(self):
+        return self.file.url
+
+    def get_upload_to(self, filename):
+        folder_name = 'video_tracks'
+        filename = self.file.field.storage.get_valid_name(filename)
+        return os.path.join(folder_name, filename)